+++ docs/source/installation.rst
@@ -1,0 +1,33 @@
+.. _installation_ref:
+
+Installation
+============
+
+PyHBorehole is supported for Python version 3.10 and younger. Previous versions are officially not supported.
+It is recommended to create a new virtual environment using the `Anaconda Distribution <https://www.anaconda.com/download>`_ before using PyBorehole.
+The main dependencies of PyBorehole are `GeoPandas <https://geopandas.org/en/stable/>`_ and `Matplotlib <https://matplotlib.org/>`_ for plotting, and more secondary dependencies like `Pandas <https://pandas.pydata.org/>`_, `NumPy <https://numpy.org/>`_, `Shapely <https://shapely.readthedocs.io/en/stable/manual.html>`_, etc.
+
+Installation via PyPi
+---------------------
+
+PyBorehole can be installed via `PyPi <https://pypi.org/>`_ using::
+
+    pip install pyborehole
+
+
+Installation via Anaconda
+--------------------------
+
+PyBorehole is also available from `conda-forge <https://conda-forge.org/>`_::
+
+    conda install -c conda-forge pyhborehole
+
+
+Forking or cloning the repository
+---------------------------------
+
+The PyHeatDemand repository can be forked or cloned from https://github.com/AlexanderJuestel/pyborehole::
+
+    git clone https://github.com/AlexanderJuestel/pyborehole.git
+
+A list of `requirements.txt <https://github.com/AlexanderJuestel/pyborehole/blob/main/requirements.txt>`_ and an `environment.yml <https://github.com/AlexanderJuestel/pyborehole/blob/main/environment.yml>`_ can be found in the repository.
